+++ runtime/src/ujima/desktop/http.clj
@@ -75,7 +75,6 @@
 
       "GET  /app/catalog"  (fn [_] {:status 200 :body {:apps (app/catalog-listing)}})
 
-      "GET  /assets/launcher/**"  (fn [{[tail] :path-params}] (files/static-file static-root "launcher" tail))
       "GET  /assets/icons/**"     (fn [{[tail] :path-params}] (files/static-file static-root "icons" tail))
       "GET  /assets/app-icon/*"   (fn [{[id]   :path-params}] (files/icon-file id))
       "GET  /assets/wall.png"     (fn [_] (files/wall static-root "wall.png"))